+++ src/Console/DirectiveCommand.php
@@ -197,7 +197,6 @@ class DirectiveCommand extends LighthouseGeneratorCommand
             null,
             true
         );
-        // @phpstan-ignore-next-line TODO remove with newer PHPStan
         assert(is_array($implementedInterfaces), 'Because we set $multiple = true');
 
         foreach ($implementedInterfaces as $interface) {
@@ -217,7 +216,6 @@ class DirectiveCommand extends LighthouseGeneratorCommand
             null,
             true
         );
-        // @phpstan-ignore-next-line TODO remove with newer PHPStan
         assert(is_array($usedLocations), 'Because we set $multiple = true');
 
         foreach ($usedLocations as $location) {